20260224 Entity, Repository, Service, Controller 전체 수정 및 View 추가

+++ src/main/java/com/example/Library/controller/RentalController.java
@@ -19,13 +19,10 @@ public class RentalController {
             @RequestParam("bookId") Long bookId) {
 
         try {
-            
-            rentalService.rentBook(userId, bookId);
-            
+            rentalService.rentBook(userId.intValue(), bookId); // Repository 파라미터 타입에 맞게 Long을 Integer로 변환
             return "🎉 성공! " + userId + "번 회원이 " + bookId + "번 책을 무사히 대출했습니다. DB를 확인해보세요!";
 
         } catch (Exception e) {
-            
             return "❌ 실패: " + e.getMessage();
         }
     }
